fix: address URL import review findings
