+++ src-tauri/Cargo.toml
@@ -29,6 +29,7 @@ chrono = { version = "0.4", features = ["serde"] }
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
 futures-util = "0.3"
 base64 = "0.22"
+once_cell = "1"
 regex = "1"
 dirs = "5"
 tauri-plugin-dialog = "2"
